+++ aa_skyhook_monitor/models/__init__.py
@@ -1,2 +1,10 @@
 # flake8: noqa
-from .skyhook import *
+from .skyhook import (
+    RaidableSkyhook,
+    RaidWatchConstellation,
+    RaidWatchRegion,
+    Skyhook,
+    SkyhookConfiguration,
+    SkyhookOwner,
+    SkyhookReagent,
+)