+++ aa_skyhook_monitor/urls.py
@@ -5,5 +5,6 @@ app_name = 'aa_skyhook_monitor'
 
 urlpatterns = [
     path('', views.index, name='index'),
+    path('raidable/', views.raidable, name='raidable'),
     path('add-owner/', views.add_owner, name='add_owner'),
 ]